requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
certifi==2024.2.2
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fire==0.5.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.1
joblib==1.3.2
matplotlib-inline==0.1.6
openai==1.12.0
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
python-dotenv==1.0.1
retry==0.9.2
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
termcolor==2.4.0
tomli==2.0.1
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13
